+++ impl/src/cmp/partial_eq.rs
@@ -12,7 +12,7 @@ use crate::utils::{
     attr::{self, ParseMultiple as _},
     pattern_matching::FieldsExt as _,
     structural_inclusion::TypeExt as _,
-    GenericsSearch, HashSet,
+    GenericsSearch, HashMap, HashSet, Spanning,
 };
 
 /// Expands a [`PartialEq`] derive macro.
@@ -33,15 +33,38 @@ pub fn expand(input: &syn::DeriveInput, _: &'static str) -> syn::Result<TokenStr
             }
             if !has_skipped_variants {
                 let mut skipped_fields = SkippedFields::default();
+                let mut custom_eq_functions = FieldsWithCustomEqFunction::default();
                 'fields: for (n, field) in data.fields.iter().enumerate() {
-                    for attr_name in [&attr_name, &secondary_attr_name] {
-                        if attr::Skip::parse_attrs(&field.attrs, attr_name)?.is_some() {
+                    match attr::WithOrSkip::parse_attrs(&field.attrs, &attr_name)? {
+                        Some(Spanning {
+                            item: attr::WithOrSkip::Skip,
+                            ..
+                        }) => {
                             _ = skipped_fields.insert(n);
                             continue 'fields;
                         }
+                        Some(Spanning {
+                            item: attr::WithOrSkip::With(with),
+                            ..
+                        }) => {
+                            custom_eq_functions.insert(n, with.func);
+                            continue 'fields;
+                        }
+                        None => {}
+                    }
+                    if attr::Skip::parse_attrs(&field.attrs, &secondary_attr_name)?
+                        .is_some()
+                    {
+                        _ = skipped_fields.insert(n);
+                        continue 'fields;
                     }
                 }
-                variants.push((None, &data.fields, skipped_fields));
+                variants.push((
+                    None,
+                    &data.fields,
+                    skipped_fields,
+                    custom_eq_functions,
+                ));
             }
         }
         syn::Data::Enum(data) => {
@@ -53,15 +76,38 @@ pub fn expand(input: &syn::DeriveInput, _: &'static str) -> syn::Result<TokenStr
                     }
                 }
                 let mut skipped_fields = SkippedFields::default();
+                let mut custom_eq_functions = FieldsWithCustomEqFunction::default();
                 'fields: for (n, field) in variant.fields.iter().enumerate() {
-                    for attr_name in [&attr_name, &secondary_attr_name] {
-                        if attr::Skip::parse_attrs(&field.attrs, attr_name)?.is_some() {
+                    match attr::WithOrSkip::parse_attrs(&field.attrs, &attr_name)? {
+                        Some(Spanning {
+                            item: attr::WithOrSkip::Skip,
+                            ..
+                        }) => {
                             _ = skipped_fields.insert(n);
                             continue 'fields;
                         }
+                        Some(Spanning {
+                            item: attr::WithOrSkip::With(with),
+                            ..
+                        }) => {
+                            custom_eq_functions.insert(n, with.func);
+                            continue 'fields;
+                        }
+                        None => {}
+                    }
+                    if attr::Skip::parse_attrs(&field.attrs, &secondary_attr_name)?
+                        .is_some()
+                    {
+                        _ = skipped_fields.insert(n);
+                        continue 'fields;
                     }
                 }
-                variants.push((Some(&variant.ident), &variant.fields, skipped_fields));
+                variants.push((
+                    Some(&variant.ident),
+                    &variant.fields,
+                    skipped_fields,
+                    custom_eq_functions,
+                ));
             }
         }
         syn::Data::Union(data) => {
@@ -84,6 +130,9 @@ pub fn expand(input: &syn::DeriveInput, _: &'static str) -> syn::Result<TokenStr
 /// Indices of [`syn::Field`]s marked with an [`attr::Skip`].
 type SkippedFields = HashSet<usize>;
 
+/// Mapping from [`syn::Field`] marked with an [`attr::With`] to the custom eq function.
+type FieldsWithCustomEqFunction = HashMap<usize, attr::Callable>;
+
 /// Expansion of a macro for generating a structural [`PartialEq`] implementation of an enum or a
 /// struct.
 struct StructuralExpansion<'i> {
@@ -93,7 +142,12 @@ struct StructuralExpansion<'i> {
     self_ty: (&'i syn::Ident, &'i syn::Generics),
 
     /// [`syn::Fields`] of the enum/struct to be compared in this [`StructuralExpansion`].
-    variants: Vec<(Option<&'i syn::Ident>, &'i syn::Fields, SkippedFields)>,
+    variants: Vec<(
+        Option<&'i syn::Ident>,
+        &'i syn::Fields,
+        SkippedFields,
+        FieldsWithCustomEqFunction,
+    )>,
 
     /// Indicator whether some original enum variants where skipped with an [`attr::Skip`].
     has_skipped_variants: bool,
@@ -144,7 +198,7 @@ impl StructuralExpansion<'_> {
         let match_arms = self
             .variants
             .iter()
-            .filter_map(|(variant, all_fields, skipped_fields)| {
+            .filter_map(|(variant, all_fields, skipped_fields, custom_eq_functions)| {
                 if all_fields.is_empty() || skipped_fields.len() == all_fields.len() {
                     return None;
                 }
@@ -160,7 +214,15 @@ impl StructuralExpansion<'_> {
                     .map(|num| {
                         let self_val = format_ident!("__self_{num}");
                         let other_val = format_ident!("__other_{num}");
-                        punctuated::Pair::Punctuated(quote! { #self_val #cmp #other_val }, &chain)
+                        let equality = custom_eq_functions
+                            .get(&num)
+                            .map(|eq_fn| {
+                                let maybe_not = (!eq).then(|| quote! {!});
+                                quote! { #maybe_not (#eq_fn)(#self_val, #other_val) }
+                            }
+                            ).unwrap_or_else(|| quote! { #self_val #cmp #other_val }
+                        );
+                        punctuated::Pair::Punctuated(equality, &chain)
                     })
                     .collect::<Punctuated<TokenStream, _>>();
                 _ = val_eqs.pop_punct();
@@ -178,14 +240,14 @@ impl StructuralExpansion<'_> {
                 });
             let unreachable_arm = (self.variants.len() > 1
                 && no_fields_arm.is_none())
-            .then(|| {
-                quote! {
+                .then(|| {
+                    quote! {
                     // SAFETY: This arm is never reachable due to `mem::discriminant()` comparison
                     //         preceding the expanded `match (self, other)` expression, but is
                     //         required by it when there is more than one variant.
                     _ => unsafe { derive_more::core::hint::unreachable_unchecked() },
                 }
-            });
+                });
 
             quote! {
                 match (self, __other) {
@@ -221,7 +283,7 @@ impl ToTokens for StructuralExpansion<'_> {
         {
             let self_ty: syn::Type = parse_quote! { Self };
             let implementor_ty: syn::Type = parse_quote! { #ty #ty_generics };
-            for (_, all_fields, skipped_fields) in &self.variants {
+            for (_, all_fields, skipped_fields, _) in &self.variants {
                 for field_ty in
                     all_fields.iter().enumerate().filter_map(|(n, field)| {
                         (!skipped_fields.contains(&n)).then_some(&field.ty)